Move non-command code out of management/commands
